task: remove old bundle files
