+++ lib/src/models/data-bundle.dart
@@ -1,16 +1,41 @@
-import 'package:json_annotation/json_annotation.dart';
+import 'dart:typed_data';
 
-import 'models.dart';
+import 'package:arweave/arweave.dart';
 
-part 'data-bundle.g.dart';
+import '../../utils.dart';
 
-@JsonSerializable()
 class DataBundle {
   final List<DataItem> items;
 
   DataBundle({required this.items});
 
-  factory DataBundle.fromJson(Map<String, dynamic> json) =>
-      _$DataBundleFromJson(json);
-  Map<String, dynamic> toJson() => _$DataBundleToJson(this);
+  Future<Uint8List> asBlob(Wallet wallet) async {
+    final headers = Uint16List(64 * items.length);
+    final binaries = await Future.wait(
+      items.map((d) async {
+        // Sign DataItem
+        var index = items.indexOf(d);
+        final id = await d.sign(wallet);
+        // Create header array
+        final header = Uint8List(64);
+        // Set offset
+        header.setAll(0, longTo32ByteArray(d.asBinary().lengthInBytes));
+        // Set id
+        header.setAll(32, id);
+        // Add header to array of headers
+        headers.setAll(64 * index, header);
+        // Convert to array for flattening
+        final raw = d.asBinary();
+        return raw.asUint8List();
+      }),
+    ).then((a) {
+      return a.reduce((a, e) => a + e as Uint8List);
+    });
+    final buffer = Uint8List.fromList([
+      ...longTo32ByteArray(items.length),
+      ...headers,
+      ...binaries,
+    ]);
+    return buffer;
+  }
 }
