+++ lib/src/models/transaction-base.dart
@@ -25,7 +25,6 @@ abstract class TransactionBase {
   Future<Uint8List> getSignatureData();
 
   Future<void> sign(Wallet wallet);
-  Future<void> signWithRawSignature(Uint8List rawSignature);
 
   Future<bool> verify();
 }
